+++ game_patch/graphics/gr_color.cpp
@@ -88,6 +88,12 @@ static_assert(static_cast<ColorChannel<5>>(ColorChannel<8>{0}).value == 0);
 static_assert(static_cast<ColorChannel<8>>(ColorChannel<0>{0}).value == 255);
 static_assert(static_cast<ColorChannel<0>>(ColorChannel<0>{0}).value == 0);
 
+template<unsigned BITS>
+struct ChannelTrait
+{
+    static constexpr unsigned bits = BITS;
+};
+
 template<unsigned BITS, unsigned OFFSET>
 struct ChannelBitOffsetTrait
 {
@@ -201,6 +207,63 @@ struct PixelFormatTrait<rf::BM_FORMAT_888_BGR>
     using Blue  = ChannelIndexTrait<8, 2>;
 };
 
+template<>
+struct PixelFormatTrait<rf::BM_FORMAT_DXT1>
+{
+    using Pixel = void;
+    using PaletteEntry = void;
+
+    using Alpha = ChannelTrait<1>;
+    using Red   = ChannelBitOffsetTrait<5, 11>;
+    using Green = ChannelBitOffsetTrait<6, 5>;
+    using Blue  = ChannelBitOffsetTrait<5, 0>;
+
+    struct Block
+    {
+        uint16_t color[2];
+        uint32_t lut;
+    };
+};
+
+template<>
+struct PixelFormatTrait<rf::BM_FORMAT_DXT3>
+{
+    using Pixel = void;
+    using PaletteEntry = void;
+
+    using Alpha = ChannelTrait<4>;
+    using Red   = ChannelBitOffsetTrait<5, 11>;
+    using Green = ChannelBitOffsetTrait<6, 5>;
+    using Blue  = ChannelBitOffsetTrait<5, 0>;
+
+    struct Block
+    {
+        uint64_t alpha_lut;
+        uint16_t color[2];
+        uint32_t lut;
+    };
+};
+
+template<>
+struct PixelFormatTrait<rf::BM_FORMAT_DXT5>
+{
+    using Pixel = void;
+    using PaletteEntry = void;
+
+    using Alpha = ChannelTrait<8>;
+    using Red   = ChannelBitOffsetTrait<5, 11>;
+    using Green = ChannelBitOffsetTrait<6, 5>;
+    using Blue  = ChannelBitOffsetTrait<5, 0>;
+
+    struct Block
+    {
+        uint8_t alpha[2];
+        uint16_t alpha_lut[3];
+        uint16_t color[2];
+        uint32_t lut;
+    };
+};
+
 template<rf::BmFormat FMT>
 struct PixelColor
 {
@@ -235,6 +298,213 @@ struct PixelColor
         return PixelColor<FMT>{a / n, r / n, g / n, b / n};
     }
 };
+
+template<typename CH, typename T>
+ColorChannel<CH::bits> unpack_color_channel(T val)
+{
+    if constexpr (CH::bits == 0) {
+        (void) val; // unused
+        return {0};
+    }
+    else if constexpr (std::is_array_v<T>) {
+        return {val[CH::index]};
+    }
+    else {
+        constexpr int mask = (1 << CH::bits) - 1;
+        return {static_cast<int>(val >> CH::offset) & mask};
+    }
+}
+
+template<rf::BmFormat FMT>
+PixelColor<FMT> unpack_color(typename PixelFormatTrait<FMT>::Pixel value)
+{
+    using PFT = PixelFormatTrait<FMT>;
+    return {
+        unpack_color_channel<typename PFT::Alpha>(value),
+        unpack_color_channel<typename PFT::Red>(value),
+        unpack_color_channel<typename PFT::Green>(value),
+        unpack_color_channel<typename PFT::Blue>(value),
+    };
+}
+
+template<rf::BmFormat>
+class BlockDecoder;
+
+template<>
+class BlockDecoder<rf::BM_FORMAT_DXT1>
+{
+    using PFT = PixelFormatTrait<rf::BM_FORMAT_DXT1>;
+    using Block = PFT::Block;
+    const Block* block_;
+
+public:
+    BlockDecoder(const Block* block) : block_(block) {}
+
+    PixelColor<rf::BM_FORMAT_DXT1> decode(int x, int y)
+    {
+        auto index = y * 4 + x;
+        auto c0_u16 = block_->color[0];
+        auto c1_u16 = block_->color[1];
+        auto c0 = unpack_color<rf::BM_FORMAT_565_RGB>(c0_u16);
+        auto c1 = unpack_color<rf::BM_FORMAT_565_RGB>(c1_u16);
+        auto lut_val = (block_->lut >> (index * 2)) & 3;
+        if (c0_u16 > c1_u16) {
+            switch (lut_val) {
+                case 0: return c0;
+                case 1: return c1;
+                case 2: return (c0 * 2 + c1) / 3;
+                case 3: return (c0 + c1 * 2) / 3;
+            }
+        }
+        else {
+            switch (lut_val) {
+                case 0: return c0;
+                case 1: return c1;
+                case 2: return (c0 + c1) / 2;
+                case 3: return {}; // transparent
+            }
+        }
+        // unreachable...
+        return {};
+    }
+};
+
+template<>
+class BlockDecoder<rf::BM_FORMAT_DXT3>
+{
+    using PFT = PixelFormatTrait<rf::BM_FORMAT_DXT3>;
+    using Block = PFT::Block;
+    const Block* block_;
+
+public:
+    BlockDecoder(const Block* block) : block_(block) {}
+
+    PixelColor<rf::BM_FORMAT_DXT3> decode(int x, int y)
+    {
+        auto index = y * 4 + x;
+        auto a = static_cast<int>((block_->alpha_lut >> (index * 4)) & 15);
+        auto c = decode_color(index);
+        return {a, c.r, c.g, c.b};
+    }
+
+private:
+    PixelColor<rf::BM_FORMAT_565_RGB> decode_color(int index)
+    {
+        auto c0 = unpack_color<rf::BM_FORMAT_565_RGB>(block_->color[0]);
+        auto c1 = unpack_color<rf::BM_FORMAT_565_RGB>(block_->color[1]);
+        auto lut_val = (block_->lut >> (index * 2)) & 3;
+        switch (lut_val) {
+            case 0: return c0;
+            case 1: return c1;
+            case 2: return (c0 * 2 + c1) / 3;
+            case 3: return (c0 + c1 * 2) / 3;
+        }
+        // unreachable...
+        return {};
+    }
+};
+
+template<>
+class BlockDecoder<rf::BM_FORMAT_DXT5>
+{
+    using PFT = PixelFormatTrait<rf::BM_FORMAT_DXT5>;
+    using Block = PFT::Block;
+    const Block* block_;
+
+public:
+    BlockDecoder(const Block* block) : block_(block) {}
+
+    PixelColor<rf::BM_FORMAT_DXT5> decode(int x, int y)
+    {
+        auto index = y * 4 + x;
+        auto a = decode_alpha(index);
+        auto c = decode_color(index);
+        return {a, c.r, c.g, c.b};
+    }
+
+private:
+    int decode_alpha(int index)
+    {
+        auto a0 = block_->alpha[0];
+        auto a1 = block_->alpha[1];
+        auto alpha_lut_64 = static_cast<uint64_t>(block_->alpha_lut[2]) << 32
+            | (static_cast<uint64_t>(block_->alpha_lut[1]) << 16)
+            | static_cast<uint64_t>(block_->alpha_lut[0]);
+        auto lut_val = (alpha_lut_64 >> (index * 3)) & 7;
+        if (a0 > a1) {
+            switch (lut_val) {
+                case 0: return a0;
+                case 1: return a1;
+                case 2: return (6 * a0 + 1 * a1) / 7;
+                case 3: return (5 * a0 + 2 * a1) / 7;
+                case 4: return (4 * a0 + 3 * a1) / 7;
+                case 5: return (3 * a0 + 4 * a1) / 7;
+                case 6: return (2 * a0 + 5 * a1) / 7;
+                case 7: return (1 * a0 + 6 * a1) / 7;
+            }
+        }
+        else {
+            switch (lut_val) {
+                case 0: return a0;
+                case 1: return a1;
+                case 2: return (4 * a0 + 1 * a1) / 5;
+                case 3: return (3 * a0 + 2 * a1) / 5;
+                case 4: return (2 * a0 + 3 * a1) / 5;
+                case 5: return (1 * a0 + 4 * a1) / 5;
+                case 6: return 0;
+                case 7: return 255;
+            }
+        }
+        // unreachable...
+        return {};
+    }
+
+    PixelColor<rf::BM_FORMAT_565_RGB> decode_color(int index)
+    {
+        auto c0 = unpack_color<rf::BM_FORMAT_565_RGB>(block_->color[0]);
+        auto c1 = unpack_color<rf::BM_FORMAT_565_RGB>(block_->color[1]);
+        auto lut_val = (block_->lut >> (index * 2)) & 3;
+        switch (lut_val) {
+            case 0: return c0;
+            case 1: return c1;
+            case 2: return (c0 * 2 + c1) / 3;
+            case 3: return (c0 + c1 * 2) / 3;
+        }
+        // unreachable...
+        return {};
+    }
+};
+
+template<rf::BmFormat Fmt>
+rf::Color DecodeBlockCompressedPixelInternal(const void* block, int x, int y)
+{
+    using PFT = PixelFormatTrait<Fmt>;
+    using Block = typename PFT::Block;
+    auto ptr = reinterpret_cast<const Block*>(block);
+    auto blk_dec = BlockDecoder<Fmt>{ptr};
+    PixelColor<rf::BmFormat::BM_FORMAT_8888_ARGB> color = blk_dec.decode(x, y);
+    return {
+        static_cast<uint8_t>(color.r.value),
+        static_cast<uint8_t>(color.g.value),
+        static_cast<uint8_t>(color.b.value),
+        static_cast<uint8_t>(color.a.value),
+    };
+}
+
+rf::Color DecodeBlockCompressedPixel(void* block, rf::BmFormat format, int x, int y)
+{
+    switch (format) {
+        case rf::BM_FORMAT_DXT1:
+            return DecodeBlockCompressedPixelInternal<rf::BM_FORMAT_DXT1>(block, x, y);
+        case rf::BM_FORMAT_DXT3:
+            return DecodeBlockCompressedPixelInternal<rf::BM_FORMAT_DXT3>(block, x, y);
+        case rf::BM_FORMAT_DXT5:
+            return DecodeBlockCompressedPixelInternal<rf::BM_FORMAT_DXT5>(block, x, y);
+        // lets skip DXT2 and DXT4 for now because they are unpopular
+        default:
+            return rf::Color{0, 0, 0, 0};
+    };
+}
 
 template<rf::BmFormat PF>
 class PixelsReader
@@ -475,10 +745,10 @@ CodeInjection LevelLoadLightmaps_color_conv_patch{
             lightmap->buf[i] = std::max(lightmap->buf[i], (uint8_t)(4 << 3)); // 32
     #endif
 
-        bool success = ConvertSurfaceFormat(lock.bits, lock.pixel_format, lightmap->buf,
-            rf::BM_FORMAT_888_BGR, lightmap->w, lightmap->h, lock.pitch, 3 * lightmap->w, nullptr);
+        bool success = ConvertSurfaceFormat(lock.data, lock.format, lightmap->buf,
+            rf::BM_FORMAT_888_BGR, lightmap->w, lightmap->h, lock.stride_in_bytes, 3 * lightmap->w, nullptr);
         if (!success)
-            xlog::error("ConvertBitmapFormat failed for lightmap (dest format %d)", lock.pixel_format);
+            xlog::error("ConvertBitmapFormat failed for lightmap (dest format %d)", lock.format);
 
         rf::GrUnlock(&lock);
     },
@@ -500,15 +770,15 @@ CodeInjection GSurface_CalculateLightmap_color_conv_patch{
         int offset_y = StructFieldRef<int>(face_light_info, 20);
         int offset_x = StructFieldRef<int>(face_light_info, 16);
         int src_width = lightmap.w;
-        int dst_pixel_size = GetBmFormatSize(lock.pixel_format);
+        int dst_pixel_size = GetBmFormatSize(lock.format);
         uint8_t* src_data = lightmap.buf + 3 * (offset_x + offset_y * src_width);
-        uint8_t* dst_data = &lock.bits[dst_pixel_size * offset_x + offset_y * lock.pitch];
+        uint8_t* dst_data = &lock.data[dst_pixel_size * offset_x + offset_y * lock.stride_in_bytes];
         int height = StructFieldRef<int>(face_light_info, 28);
         int src_pitch = 3 * src_width;
-        bool success = ConvertSurfaceFormat(dst_data, lock.pixel_format, src_data, rf::BM_FORMAT_888_BGR,
-            src_width, height, lock.pitch, src_pitch);
+        bool success = ConvertSurfaceFormat(dst_data, lock.format, src_data, rf::BM_FORMAT_888_BGR,
+            src_width, height, lock.stride_in_bytes, src_pitch);
         if (!success)
-            xlog::error("ConvertSurfaceFormat failed for geomod (fmt %d)", lock.pixel_format);
+            xlog::error("ConvertSurfaceFormat failed for geomod (fmt %d)", lock.format);
         rf::GrUnlock(&lock);
     },
 };
@@ -533,13 +803,13 @@ CodeInjection GSurface_AllocLightmap_color_conv_patch{
         int src_offset = 3 * (offset_x + src_width * StructFieldRef<int>(face_light_info, 20)); // src offset
         uint8_t* src_data = src_offset + src_data_begin;
         int height = StructFieldRef<int>(face_light_info, 28);
-        int dst_pixel_size = GetBmFormatSize(lock.pixel_format);
-        uint8_t* dst_row_ptr = &lock.bits[dst_pixel_size * offset_x + offset_y * lock.pitch];
+        int dst_pixel_size = GetBmFormatSize(lock.format);
+        uint8_t* dst_row_ptr = &lock.data[dst_pixel_size * offset_x + offset_y * lock.stride_in_bytes];
         int src_pitch = 3 * src_width;
-        bool success = ConvertSurfaceFormat(dst_row_ptr, lock.pixel_format, src_data, rf::BM_FORMAT_888_BGR,
-                                                 src_width, height, lock.pitch, src_pitch);
+        bool success = ConvertSurfaceFormat(dst_row_ptr, lock.format, src_data, rf::BM_FORMAT_888_BGR,
+                                                 src_width, height, lock.stride_in_bytes, src_pitch);
         if (!success)
-            xlog::error("ConvertBitmapFormat failed for geomod2 (fmt %d)", lock.pixel_format);
+            xlog::error("ConvertBitmapFormat failed for geomod2 (fmt %d)", lock.format);
         rf::GrUnlock(&lock);
     },
 };
@@ -563,31 +833,31 @@ CodeInjection GProcTexUpdateWater_patch{
         }
 
         try {
-            CallWithPixelFormat(src_lock_data.pixel_format, [=](auto s) {
-                CallWithPixelFormat(dst_lock_data.pixel_format, [=](auto d) {
+            CallWithPixelFormat(src_lock_data.format, [=](auto s) {
+                CallWithPixelFormat(dst_lock_data.format, [=](auto d) {
                     auto& byte_1370f90 = AddrAsRef<uint8_t[256]>(0x1370F90);
                     auto& byte_1371b14 = AddrAsRef<uint8_t[256]>(0x1371B14);
                     auto& byte_1371090 = AddrAsRef<uint8_t[512]>(0x1371090);
 
-                    uint8_t* dst_row_ptr = dst_lock_data.bits;
-                    int src_pixel_size = GetBmFormatSize(src_lock_data.pixel_format);
+                    uint8_t* dst_row_ptr = dst_lock_data.data;
+                    int src_pixel_size = GetBmFormatSize(src_lock_data.format);
 
-                    for (int y = 0; y < dst_lock_data.height; ++y) {
+                    for (int y = 0; y < dst_lock_data.h; ++y) {
                         int t1 = byte_1370f90[y];
                         int t2 = byte_1371b14[y];
                         uint8_t* off_arr = &byte_1371090[-t1];
                         PixelsWriter<decltype(d)::value> wrt{dst_row_ptr};
-                        for (int x = 0; x < dst_lock_data.width; ++x) {
+                        for (int x = 0; x < dst_lock_data.w; ++x) {
                             int src_x = t1;
                             int src_y = t2 + off_arr[t1];
-                            int src_x_limited = src_x & (dst_lock_data.width - 1);
-                            int src_y_limited = src_y & (dst_lock_data.height - 1);
-                            const uint8_t* src_ptr = src_lock_data.bits + src_x_limited * src_pixel_size + src_y_limited * src_lock_data.pitch;
+                            int src_x_limited = src_x & (dst_lock_data.w - 1);
+                            int src_y_limited = src_y & (dst_lock_data.h - 1);
+                            const uint8_t* src_ptr = src_lock_data.data + src_x_limited * src_pixel_size + src_y_limited * src_lock_data.stride_in_bytes;
                             PixelsReader<decltype(s)::value> rdr{src_ptr};
                             wrt.write(rdr.read());
                             ++t1;
                         }
-                        dst_row_ptr += dst_lock_data.pitch;
+                        dst_row_ptr += dst_lock_data.stride_in_bytes;
                     }
                 });
             });
@@ -646,7 +916,7 @@ CodeInjection MonitorUpdateStatic_patch{
 
         auto& lock = *reinterpret_cast<rf::GrLockInfo*>(regs.esp + 0x2C - 0x20);
         auto pixel_ptr = reinterpret_cast<char*>(regs.esi);
-        int bytes_per_pixel = GetBmFormatSize(lock.pixel_format);
+        int bytes_per_pixel = GetBmFormatSize(lock.format);
 
         std::fill(pixel_ptr, pixel_ptr + bytes_per_pixel, white ? '\0' : '\xFF');
         regs.esi += bytes_per_pixel;
@@ -659,7 +929,7 @@ CodeInjection MonitorUpdateOff_patch{
     0x00412430,
     [](auto& regs) {
         auto& lock = *reinterpret_cast<rf::GrLockInfo*>(regs.esp + 0x34 - 0x20);
-        regs.ecx = lock.height * lock.pitch;
+        regs.ecx = lock.h * lock.stride_in_bytes;
         regs.eip = 0x0041243F;
     },
 };