+++ game_patch/graphics/gr_texture.cpp
@@ -208,39 +208,75 @@ CodeInjection gr_d3d_create_vram_texture_with_mipmaps_pitch_fix{
     },
 };
 
+int GetBytesPerCompressedBlock(rf::BmFormat format)
+{
+    switch (format) {
+        case rf::BM_FORMAT_DXT1:
+            return 8;
+        case rf::BM_FORMAT_DXT2:
+        case rf::BM_FORMAT_DXT3:
+        case rf::BM_FORMAT_DXT4:
+        case rf::BM_FORMAT_DXT5:
+            return 16;
+        default:
+            assert(false);
+            return 0;
+    }
+}
+
 FunHook <void(int, float, float, rf::Color*)> gr_d3d_get_texel_hook{
     0x0055CFA0,
-    [](int bmh, float u, float v, rf::Color* out_color) {
-        if (BmIsCompressedFormat(rf::BmGetFormat(bmh))) {
-            // Reading pixels is not yet supported in compressed texture formats
-            // This is only used when shooting at a texture with alpha
-            out_color->red = 255;
-            out_color->green = 255;
-            out_color->blue = 255;
-            out_color->alpha = 255;
+    [](int bm_handle, float u, float v, rf::Color* out_color) {
+        if (BmIsCompressedFormat(rf::BmGetFormat(bm_handle))) {
+            // This function is only used when shooting at a texture with alpha
+            rf::GrLockInfo lock;
+            if (rf::GrLock(bm_handle, 0, &lock, 0)) {
+                // Make sure u and v are in [0, 1] range
+                // Assume wrap addressing mode
+                u = std::fmod(u, 1.0f);
+                v = std::fmod(v, 1.0f);
+                if (u < 0.0f) {
+                    u += 1.0f;
+                }
+                if (v < 0.0f) {
+                    v += 1.0f;
+                }
+                int x = static_cast<int>(u * lock.w + 0.5f);
+                int y = static_cast<int>(v * lock.h + 0.5f);
+                constexpr int block_w = 4;
+                constexpr int block_h = 4;
+                auto bytes_per_block = GetBytesPerCompressedBlock(lock.format);
+                auto block = lock.data + (y / block_h) * lock.stride_in_bytes + x / block_w * bytes_per_block;
+                *out_color = DecodeBlockCompressedPixel(block, lock.format, x % block_w, y % block_h);
+                rf::GrUnlock(&lock);
+            }
+            else {
+                out_color->Set(255, 255, 255, 255);
+            }
         }
         else {
-            gr_d3d_get_texel_hook.CallTarget(bmh, u, v, out_color);
+            gr_d3d_get_texel_hook.CallTarget(bm_handle, u, v, out_color);
         }
     },
 };
 
-FunHook<int(int, rf::GrD3DTexture&)> gr_d3d_create_texture_hook{
-    0x0055CC00,
-    [](int bm_handle, rf::GrD3DTexture& tslot) {
-        auto result = gr_d3d_create_texture_hook.CallTarget(bm_handle, tslot);
-        if (result != 1) {
-            xlog::warn("Failed to load texture '%s'", rf::BmGetFilename(bm_handle));
-            // Note: callers of this function expects zero result on failure
-            return 0;
-        }
-        auto pixel_fmt = rf::BmGetFormat(bm_handle);
-        if (pixel_fmt == rf::BM_FORMAT_RENDER_TARGET) {
-            g_default_pool_tslots.insert(&tslot);
-        }
-        return result;
-    },
-};
+extern FunHook<int(int, rf::GrD3DTexture&)> gr_d3d_create_texture_hook;
+
+int gr_d3d_create_texture(int bm_handle, rf::GrD3DTexture& tslot) {
+    auto result = gr_d3d_create_texture_hook.CallTarget(bm_handle, tslot);
+    if (result != 1) {
+        xlog::warn("Failed to load texture '%s'", rf::BmGetFilename(bm_handle));
+        // Note: callers of this function expects zero result on failure
+        return 0;
+    }
+    auto pixel_fmt = rf::BmGetFormat(bm_handle);
+    if (pixel_fmt == rf::BM_FORMAT_RENDER_TARGET) {
+        g_default_pool_tslots.insert(&tslot);
+    }
+    return result;
+}
+
+FunHook<int(int, rf::GrD3DTexture&)> gr_d3d_create_texture_hook{0x0055CC00, gr_d3d_create_texture};
 
 FunHook<void(rf::GrD3DTexture&)> gr_d3d_free_texture_hook{
     0x0055B640,
@@ -256,6 +292,36 @@ FunHook<void(rf::GrD3DTexture&)> gr_d3d_free_texture_hook{
         }
     },
 };
+
+bool gr_d3d_lock(int bm_handle, int section, rf::GrLockInfo *lock) {
+    auto& tslot = rf::gr_d3d_textures[rf::BmHandleToIdxAnimAware(bm_handle)];
+    if (tslot.num_sections < 1 || tslot.bm_handle != bm_handle) {
+        auto ret = gr_d3d_create_texture(bm_handle, tslot);
+        if (ret != 1) {
+            return false;
+        }
+    }
+    D3DLOCKED_RECT locked_rect;
+    auto d3d_texture = tslot.sections[section].d3d_texture;
+    if (FAILED(d3d_texture->LockRect(0, &locked_rect, nullptr, 0))) {
+        return false;
+    }
+
+    tslot.lock_count++;
+    D3DSURFACE_DESC desc;
+    d3d_texture->GetLevelDesc(0, &desc);
+
+    lock->data = reinterpret_cast<rf::ubyte*>(locked_rect.pBits);
+    lock->stride_in_bytes = locked_rect.Pitch;
+    lock->w = desc.Width;
+    lock->h = desc.Height;
+    lock->section = section;
+    lock->bm_handle = bm_handle;
+    lock->format = GetBmFormatFromD3DFormat(desc.Format);
+    return true;
+}
+
+static FunHook<bool(int, int, rf::GrLockInfo *)> gr_d3d_lock_hook{0x0055CE00, gr_d3d_lock};
 
 // FunHook<void(int, int, int)> gr_d3d_set_state_and_texture_hook{
 //     0x00550850,
@@ -297,6 +363,7 @@ void ApplyTexturePatches()
     gr_d3d_create_texture_hook.Install();
     gr_d3d_free_texture_hook.Install();
     //gr_d3d_set_state_and_texture_hook.Install();
+    gr_d3d_lock_hook.Install();
 }
 
 void ReleaseAllDefaultPoolTextures()