+++ game_patch/misc/main_menu.cpp
@@ -91,7 +91,7 @@ int LoadEasterEggImage()
     if (!rf::GrLock(hbm, 0, &lock, 1))
         return -1;
 
-    rf::BmConvertFormat(lock.bits, lock.pixel_format, res_data, rf::BM_FORMAT_8888_ARGB,
+    rf::BmConvertFormat(lock.data, lock.format, res_data, rf::BM_FORMAT_8888_ARGB,
                         easter_egg_size * easter_egg_size);
     rf::GrUnlock(&lock);