+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Version 1.5.1 (not released yet)
 - add links to Red Faction Wiki in RED Help menu replacing old "Help Topics" menu item that did not work on modern
   Windows versions because of the lack of support for WinHelp format
 - fix F4 key (Maximize active viewport) in RED for screens larger than 1024x768
+- fix shooting through transparent sections of DDS textures
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------